requirements.txt
requests
tqdm
colorama
PyJWT